+++ src/components/BulkEditList/BulkEditListFilters/BulkEditListFilters.js
@@ -16,8 +16,11 @@ import { ListSelect } from './ListSelect/ListSelect';
 import { QueryTextArea } from './QueryTextArea/QueryTextArea';
 import { ListFileUploader } from '../../ListFileUploader';
 import { buildCheckboxFilterOptions } from './utils/optionsRecordIdentifiers';
-import { EDIT_CAPABILITIES } from '../../../constants/optionsRecordIdentifiers';
-import { BULK_EDIT_IDENTIFIERS } from '../../../constants/constants';
+import {
+  BULK_EDIT_IDENTIFIERS,
+  EDIT_CAPABILITIES,
+  BULK_EDIT_QUERY,
+} from '../../../constants';
 import { getFileInfo } from './utils/getFileInfo';
 import { useJobCommand, useFileUploadComand } from '../../../API/useFileUpload';
 
@@ -25,6 +28,7 @@ export const BulkEditListFilters = ({
   setFileUploadedName,
   isFileUploaded,
   setIsFileUploaded,
+  setCountOfRecords,
 }) => {
   const [isDropZoneActive, setDropZoneActive] = useState(false);
   const [fileExtensionModalOpen, setFileExtensionModalOpen] = useState(false);
@@ -101,11 +105,15 @@ export const BulkEditListFilters = ({
     try {
       const { id } = await requestJobId({ recordIdentifier, editType: BULK_EDIT_IDENTIFIERS });
 
-      await fileUpload({ id, fileToUpload });
+      const recordsCount = await fileUpload({ id, fileToUpload });
+
+      setCountOfRecords(recordsCount);
+
+      const locationParams = new URLSearchParams(location.search).delete('queryText');
 
       history.replace({
         pathname: `/bulk-edit/${id}`,
-        search: buildSearch({ fileName: fileToUpload.name }, location.search),
+        search: buildSearch({ fileName: fileToUpload.name }, locationParams),
       });
 
       setIsFileUploaded(true);
@@ -126,6 +134,21 @@ export const BulkEditListFilters = ({
     } else {
       showFileExtensionModal();
     }
+  };
+
+  const handleQuerySearch = async () => {
+    const { id } = await requestJobId({
+      recordIdentifier,
+      editType: BULK_EDIT_QUERY,
+      specificParameters: { query: queryText },
+    });
+
+    const locationParams = new URLSearchParams(location.search).delete('fileName');
+
+    history.replace({
+      pathname: `/bulk-edit/${id}`,
+      search: buildSearch({ queryText }, locationParams),
+    });
   };
 
   const renderBadge = () => <Badge data-testid="filter-badge">0</Badge>;
@@ -160,7 +183,13 @@ export const BulkEditListFilters = ({
       <ButtonGroup fullWidth>
         {renderTopButtons()}
       </ButtonGroup>
-      {criteria === 'query' && <QueryTextArea queryText={queryText} setQueryText={setFilters} />}
+      {criteria === 'query' && (
+        <QueryTextArea
+          queryText={queryText}
+          setQueryText={setFilters}
+          handleQuerySearch={handleQuerySearch}
+        />
+      )}
       {criteria === 'identifier' &&
       <>
         <ListSelect
@@ -208,4 +237,5 @@ BulkEditListFilters.propTypes = {
   setFileUploadedName: PropTypes.func.isRequired,
   isFileUploaded: PropTypes.bool.isRequired,
   setIsFileUploaded: PropTypes.func.isRequired,
+  setCountOfRecords: PropTypes.func.isRequired,
 };