+++ src/components/BulkEditList/BulkEditListResult/Preview/Preview.test.js
@@ -1,18 +1,22 @@
 import { render, screen } from '@testing-library/react';
+import { MemoryRouter } from 'react-router';
 import { useOkapiKy } from '@folio/stripes/core';
 
 import '../../../../../test/jest/__mock__';
 
-import { MemoryRouter } from 'react-router';
 import { QueryClientProvider } from 'react-query';
 import { Preview } from './Preview';
 import { queryClient } from '../../../../../test/jest/utils/queryClient';
 
+jest.mock('./PreviewAccordion', () => ({
+  PreviewAccordion: () => 'PreviewAccordion',
+}));
+
 const renderPreview = () => {
   render(
-    <MemoryRouter initialEntries={['/bulk-edit/1']}>
+    <MemoryRouter initialEntries={['/bulk-edit/1?queryText=patronGroup%3D%3D"1"']}>
       <QueryClientProvider client={queryClient}>
-        <Preview fileUploadedName="Mock.csv" />,
+        <Preview />
       </QueryClientProvider>
     </MemoryRouter>,
   );
@@ -37,6 +41,7 @@ describe('Preview', () => {
   it('displays Bulk edit', () => {
     renderPreview();
 
-    expect(screen.getByText('FileName: Mock.csv')).toBeVisible();
+    expect(screen.getByText(/preview.query.title/)).toBeVisible();
+    expect(screen.getByText('PreviewAccordion')).toBeVisible();
   });
 });