+++ test/jest/utils/fileUpload.js
@@ -1,5 +1,4 @@
 import { fireEvent } from '@testing-library/react';
-import { QueryClient } from 'react-query';
 
 export function mockData(files) {
   return {